refactor(step3): 생성자 private하게 만들고 생성하는 곳을 옮기기
