--- src/main/java/step3/RacingCarInfo.java
@@ -1,13 +1,0 @@
-package step3;
-
-import java.util.List;
-
-public class RacingCarInfo {
-    public final List<Car> cars;
-    public final Integer tryCount;
-
-    public RacingCarInfo(List<Car> cars, Integer tryCount) {
-        this.cars = cars;
-        this.tryCount = tryCount;
-    }
-}
